% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse.test.R
\name{vse.test}
\alias{vse.test}
\title{Variance Scale Exponent Test}
\usage{
vse.test(x, m = 0.5)
}
\arguments{
\item{x}{A time series vector.}

\item{m}{A parameter to control the number of scales. Default is 0.5.}
}
\value{
A list with class "vsetest" containing the following components:
\item{SLmemory}{the test statistic of the Short-Long memory test.}
\item{df}{the degrees of freedom of the test.}
\item{p.value}{the p-value of the test.}
}
\description{
Use the variance scale exponent method to construct a hypothesis test about long memory in time series.
}
\examples{
set.seed(123)
x <- rnorm(1024)
vse.test(x)
}
\references{
Fu, H., Chen, W., & He, X.-J. (2018). On a class of estimation and test for long memory. In Physica A: Statistical Mechanics and its Applications (Vol. 509, pp. 906–920). Elsevier BV. https://doi.org/10.1016/j.physa.2018.06.092
}
