% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvse.R
\name{wvse}
\alias{wvse}
\title{Calculate weak variance scaling exponents}
\usage{
wvse(R)
}
\arguments{
\item{R}{A time series vector.}
}
\value{
The weak variance scaling exponent.
}
\description{
Calculate weak variance scaling exponents
}
\examples{
x <- rnorm(1024)
wvse(x)
}
