% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse.R
\name{vse}
\alias{vse}
\title{Calculate variance scaling exponents}
\usage{
vse(x, m = 0.5)
}
\arguments{
\item{x}{A time series vector.}

\item{m}{A parameter to control the number of scales. Default is 0.5.}
}
\value{
The weak variance scaling exponent.
}
\description{
Calculate variance scaling exponents
}
\examples{
x <- rnorm(1024)
vse(x)
}
