% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse.R
\name{vse}
\alias{vse}
\title{Variance Scale Exponent}
\usage{
vse(x, m = 0.5)
}
\arguments{
\item{x}{A time series vector.}

\item{m}{A parameter to control the number of scales. Default is 0.5.}
}
\value{
The variance scaling exponent.
}
\description{
Calculate the variance scaling exponent of a time series.
}
\examples{
set.seed(123)
x <- rnorm(1024)
vse(x)
}
\references{
Fu, H., Chen, W., & He, X.-J. (2018). On a class of estimation and test for long memory. In Physica A: Statistical Mechanics and its Applications (Vol. 509, pp. 906–920). Elsevier BV. https://doi.org/10.1016/j.physa.2018.06.092
}
