% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse.R
\name{vse}
\alias{vse}
\title{Variance Scale Exponent}
\usage{
vse(x, m = 0.5)
}
\arguments{
\item{x}{A time series vector.}

\item{m}{A parameter to control the number of scales. Default is 0.5.}
}
\value{
A list containing the variance scaling exponent.
}
\description{
Calculate the variance scaling exponent of a time series.
}
\examples{
set.seed(123)
x <- rnorm(1024)
vse(x)
}
