% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse.test.R
\name{vse.test}
\alias{vse.test}
\title{Variance Scale Exponent Test}
\usage{
vse.test(x, m = 0.5)
}
\arguments{
\item{x}{A time series vector.}

\item{m}{A parameter to control the number of scales. Default is 0.5.}
}
\value{
A list containing the statistic, degree of freedom, and p-value.
}
\description{
Use the variance scale exponent method to construct a hypothesis test about long memory in time series.
}
\examples{
set.seed(123)
x <- rnorm(1024)
vse.test(x)
}
