% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse4ts-package.R
\docType{package}
\name{vse4ts-package}
\alias{vse4ts}
\alias{vse4ts-package}
\title{vse4ts: Identify Memory Patterns in Time Series Using Variance Scale Exponent}
\description{
Constructing variance scale exponent, it achieves the identification of memory patterns in time series, especially for long memory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://z-my-cn.github.io/vse4ts/}
  \item Report bugs at \url{https://github.com/z-my-cn/vse4ts/issues}
}

}
\author{
\strong{Maintainer}: Mengyang Zheng \email{mengyang.zheng@outlook.com}

Authors:
\itemize{
  \item Hui Fu \email{hui_fu@hotmail.com}
}

}
\keyword{internal}
