% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vse4ts-package.R
\docType{package}
\name{vse4ts-package}
\alias{vse4ts}
\alias{vse4ts-package}
\title{vse4ts: Identify Memory Patterns in Time Series Using Variance Scale Exponent}
\description{
Methods for calculating the variance scale exponent to identify memory patterns in time series data. Includes tests for white noise, short memory, and long memory. See Fu, H. et al. (2018) \doi{10.1016/j.physa.2018.06.092}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://z-my-cn.github.io/vse4ts/}
  \item Report bugs at \url{https://github.com/z-my-cn/vse4ts/issues}
}

}
\author{
\strong{Maintainer}: Mengyang Zheng \email{mengyang.zheng@outlook.com}

Authors:
\itemize{
  \item Hui Fu \email{hui_fu@hotmail.com}
}

}
\keyword{internal}
